requirements.txt
rich
stegano
piexif
PIL
request